Refactored User class
